# handlers/products..go
package handlers

import (
	"context"
	"fmt"
	"log"
	"net/http"
	"github.com/nicholasjackson/building-microservices-youtube/product-api/data"
)

// Products is a http.Handler
type Products struct {
	l *log.Logger
}
// A list of products returns in the productsResponse
// swagger: response productsResponse
// type productsResponse struct {
// 	// All products in the system
//    	// in: body
// 	Body []data.Product
// }

//swagger:response noContent
type productsResponseWrapper struct{

}
//swagger:parameters  deleteProduct
type productIDParameterWrapper struct {
	// The id of the product to be deleted from the database 
	// in: path
	// required: true
	ID int `json:"id"`
}



// NewProducts creates a products handler with the given logger
func NewProducts(l *log.Logger) *Products {
	return &Products{l}
}

//KeyProduct struct
type KeyProduct struct{}

//MiddlewareValidateProduct func
func (p Products) MiddlewareValidateProduct(next http.Handler) http.Handler {
	return http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
		prod := data.Product{}

		err := prod.FromJSON(r.Body)
		if err != nil {
			p.l.Println("[ERROR] deserializing product", err)
			http.Error(rw, "Error reading product", http.StatusBadRequest)
			return
		}

		//validate the product
		err = prod.Validator()
		if err != nil {
			p.l.Println("[ERROR] validating product", err)
			http.Error(rw, fmt.Sprintf("Error validating product : %s", err ),http.StatusBadRequest)
			return
		}

		// add the product to the context
		ctx := context.WithValue(r.Context(), KeyProduct{}, prod)
		r = r.WithContext(ctx)

		// Call the next handler, which can be another middleware in the chain, or the final handler.
		next.ServeHTTP(rw, r)
	})
}